refactor : CopyDropdown을 ModalDropdown으로 변경
